+++ src/test/java/org/mattlang/jc/engine/tt/TTCacheTest.java
@@ -3,6 +3,7 @@ package org.mattlang.jc.engine.tt;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.Test;
+import org.mattlang.attic.tt.TTCache;
 import org.mattlang.jc.board.BoardRepresentation;
 import org.mattlang.jc.board.Color;
 import org.mattlang.jc.board.bitboard.BitBoard;