+++ src/test/java/org/mattlang/jc/perftests/PerfTTTests.java
@@ -8,10 +8,10 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 
 import org.junit.Test;
+import org.mattlang.attic.tt.TTCache;
 import org.mattlang.jc.StopWatch;
 import org.mattlang.jc.board.GameState;
 import org.mattlang.jc.board.bitboard.BitBoard;
-import org.mattlang.jc.engine.tt.TTCache;
 import org.mattlang.jc.engine.tt.TTCacheInterface;
 import org.mattlang.jc.engine.tt.TTResult;
 import org.mattlang.jc.movegenerator.MoveGenerator;